Filter System, Edit, Delete, Redirection, Register
Fixed the filter system not working with cookies, Edit and Delete now only works with users who made the submission under their account ID, Users who are logged in are now redirected to Dashboard instead of the home page. There is now also a register.ejs page for whenever registering needs to happen. I dont really see a purpose but we need it as a template. Will work more on profile, settings, and dashboard.