CSS Styling, Database Finishing Touches
More will be added to the Database for the Dashboard but it is currently finished. We are now working on CSS Styling. Styling has been finished for every page but it is still in draft form, and each page will be formatted to fit inside of the dashboard, which still needs to be fully made.